refactor : 전역 함수 store 이동
