+++ api/index.js
@@ -7,3 +7,11 @@ export const fetchData = async () => {
     console.error(error);
   }
 };
+
+export const getLocalStorage = key => {
+  return JSON.parse(localStorage.getItem(key)) ?? null;
+};
+
+export const setLocalStorage = (key, value) => {
+  localStorage.setItem(key, JSON.stringify(value));
+};
